+++ src/DependencyModules.SourceGenerator/SourceGenerator.cs
@@ -10,6 +10,7 @@ public class SourceGenerator : BaseSourceGenerator {
 
     protected override IEnumerable<IDependencyModuleSourceGenerator> AttributeSourceGenerators() {
         yield return new ServiceSourceGenerator();
+        yield return new DecoratorSourceGenerator();
     }
 
     protected override void SetupRootGenerator(IncrementalGeneratorInitializationContext context,